Change data type from String to Uri
The Images in the PhotoAdapter are saved as a MutableList of Uri; methods for adding, removing and restoring this list relates with Uris.

The ItemHolder is a separate class which is in charge of binding the uri into the views using Glide.

An extension method for MutableList called stringify is used to convert a MutableList of Uri to a List of Strings.

The SquaredRelativeLayout is moved to the common directory.

The AddButtonHolder and ItemHolder are moved to viewholder directory.

Config.kt help manage all the configurations obtained by the attributes automatically extracting this from AttributesSet.